+++ gen/state.js
@@ -1227,6 +1227,7 @@ content: () => {return setTimeout(this.success, 2000)},
         states.push(new BotInfoState(null, this, sm));
         states.push(new BotMapState(null, this, sm));
         states.push(new BotSubscriberListState(null, this, sm));
+        states.push(new BotEditPhotoState(null, this, sm));
     let transition = [];
         transition.push({
          event: "createMessageContainer", 
@@ -1365,12 +1366,20 @@ content: () => {return setTimeout(this.success, 2000)},
         
         });
         transition.push({
-            event: "botSubscriberList",
-            type: "internal",
-            mode: "push",
-
-            target: "BotSubscriberList",
-
+         event: "botSubscriberList", 
+         type: "internal", 
+         mode: "push", 
+        
+         target:"BotSubscriberList", 
+        
+        });
+        transition.push({
+         event: "botEditPhoto", 
+         type: "internal", 
+         mode: "push", 
+        
+         target:"BotEditPhoto", 
+        
         });
 
     this.states = states;
@@ -1445,7 +1454,10 @@ content: () => {return profileStore.connect(this.model.user, this.model.password
         this.handle("botMap", data);
         };
         botSubscriberList = (data) => {
-            this.handle("botSubscriberList", data);
+        this.handle("botSubscriberList", data);
+        };
+        botEditPhoto = (data) => {
+        this.handle("botEditPhoto", data);
         };
     }
     export class CubeBarState extends State {
@@ -1785,9 +1797,7 @@ content: () => {return profileStore.connect(this.model.user, this.model.password
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.promise({
-                $line: '124',
+    this.onentry = _event => { this.sm.promise({$line: '125',
 $column: '17',
 $type: 'promise',
 content: () => {return {item: messageStore.createChat(_event.data).id}}, 
@@ -2231,10 +2241,8 @@ content: () => {return {item: messageStore.createChat(_event.data).id}},
         
         
         
-         target:"PromoScene",
-            ontransition: _event => {
-                this.sm.script({
-                    $line: '151',
+         target:"PromoScene", 
+        ontransition:_event => { this.sm.script({$line: '152',
 $column: '16',
 $type: 'script',
 content: () => {return profileStore.logout(_event.data)}, 
@@ -2843,10 +2851,8 @@ content: () => {return profileStore.logout(_event.data)},
         
         
         
-         target:"CreatePrivateChat",
-            ontransition: _event => {
-                this.sm.script({
-                    $line: '179',
+         target:"CreatePrivateChat", 
+        ontransition:_event => { this.sm.script({$line: '180',
 $column: '15',
 $type: 'script',
 content: () => {return searchStore.clear()}, 
@@ -2921,9 +2927,7 @@ content: () => {return searchStore.clear()},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '184',
+    this.onentry = _event => { this.sm.script({$line: '185',
 $column: '15',
 $type: 'script',
 content: () => {return searchStore[_event.name](_event.data)}, 
@@ -2988,10 +2992,7 @@ content: () => {return searchStore[_event.name](_event.data)},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.initial = 'ProfileDetails';
-        this.onexit = _event => {
-            this.sm.script({
-                $line: '191',
+    this.initial = 'ProfileDetails'; this.onexit = _event => { this.sm.script({$line: '192',
 $column: '14',
 $type: 'script',
 content: () => {return this.shouldPop = true}, 
@@ -3145,9 +3146,7 @@ content: () => {return this.shouldPop = true},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '200',
+    this.onentry = _event => { this.sm.script({$line: '201',
 $column: '15',
 $type: 'script',
 content: () => {return profileStore.hidePosts(_event.data)}, 
@@ -3217,9 +3216,7 @@ content: () => {return profileStore.hidePosts(_event.data)},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '206',
+    this.onentry = _event => { this.sm.script({$line: '207',
 $column: '15',
 $type: 'script',
 content: () => {return profileStore.showPosts(_event.data)}, 
@@ -3770,11 +3767,20 @@ content: () => {return profileStore.showPosts(_event.data)},
         
         });
         transition.push({
-            event: "subscribers",
-
-
-            target: "BotSubscriberList",
-
+         event: "subscribers", 
+        
+        
+        
+         target:"BotSubscriberList", 
+        
+        });
+        transition.push({
+         event: "editPhoto", 
+        
+        
+        
+         target:"BotEditPhoto", 
+        
         });
 
     this.states = states;
@@ -3802,7 +3808,10 @@ content: () => {return profileStore.showPosts(_event.data)},
         this.handle("editPhotos", data);
         };
         subscribers = (data) => {
-            this.handle("subscribers", data);
+        this.handle("subscribers", data);
+        };
+        editPhoto = (data) => {
+        this.handle("editPhoto", data);
         };
     }
     export class BotPhotoListState extends State {
@@ -4341,97 +4350,32 @@ content: () => {return profileStore.showPosts(_event.data)},
 
     }
     }
-export class BotSubscriberListState extends State {
-    get storage() {
-        return this.parent.storage
-    };
+    export class BotSubscriberListState extends State {
+        get storage() { return this.parent.storage };
+        set storage(value) { this.parent.storage = value };
+        get xmppStore() { return this.parent.xmppStore };
+        set xmppStore(value) { this.parent.xmppStore = value };
+        get friendStore() { return this.parent.friendStore };
+        set friendStore(value) { this.parent.friendStore = value };
+        get profileStore() { return this.parent.profileStore };
+        set profileStore(value) { this.parent.profileStore = value };
+        get messageStore() { return this.parent.messageStore };
+        set messageStore(value) { this.parent.messageStore = value };
+        get searchStore() { return this.parent.searchStore };
+        set searchStore(value) { this.parent.searchStore = value };
+        get eventStore() { return this.parent.eventStore };
+        set eventStore(value) { this.parent.eventStore = value };
+        get model() { return this.parent.model };
+        set model(value) { this.parent.model = value };
+        get location() { return this.parent.location };
+        set location(value) { this.parent.location = value };
+        get pushStore() { return this.parent.pushStore };
+        set pushStore(value) { this.parent.pushStore = value };
+        get botStore() { return this.parent.botStore };
+        set botStore(value) { this.parent.botStore = value };
 
-    set storage(value) {
-        this.parent.storage = value
-    };
-
-    get xmppStore() {
-        return this.parent.xmppStore
-    };
-
-    set xmppStore(value) {
-        this.parent.xmppStore = value
-    };
-
-    get friendStore() {
-        return this.parent.friendStore
-    };
-
-    set friendStore(value) {
-        this.parent.friendStore = value
-    };
-
-    get profileStore() {
-        return this.parent.profileStore
-    };
-
-    set profileStore(value) {
-        this.parent.profileStore = value
-    };
-
-    get messageStore() {
-        return this.parent.messageStore
-    };
-
-    set messageStore(value) {
-        this.parent.messageStore = value
-    };
-
-    get searchStore() {
-        return this.parent.searchStore
-    };
-
-    set searchStore(value) {
-        this.parent.searchStore = value
-    };
-
-    get eventStore() {
-        return this.parent.eventStore
-    };
-
-    set eventStore(value) {
-        this.parent.eventStore = value
-    };
-
-    get model() {
-        return this.parent.model
-    };
-
-    set model(value) {
-        this.parent.model = value
-    };
-
-    get location() {
-        return this.parent.location
-    };
-
-    set location(value) {
-        this.parent.location = value
-    };
-
-    get pushStore() {
-        return this.parent.pushStore
-    };
-
-    set pushStore(value) {
-        this.parent.pushStore = value
-    };
-
-    get botStore() {
-        return this.parent.botStore
-    };
-
-    set botStore(value) {
-        this.parent.botStore = value
-    };
-
-    constructor(_, parent, sm) {
-        super({id: "BotSubscriberList"}, parent, sm);
+    constructor(_, parent, sm){
+    super({ id: "BotSubscriberList"}, parent, sm);
         const storage = this.storage;
         const xmppStore = this.xmppStore;
         const friendStore = this.friendStore;
@@ -4444,19 +4388,73 @@ export class BotSubscriberListState extends State {
         const pushStore = this.pushStore;
         const botStore = this.botStore;
 
-        let states = [];
-        let transition = [];
+    let states = [];
+    let transition = [];
 
-        this.states = states;
-        this.transitions = transition.map(el => new Transition(this, el));
+    this.states = states;
+    this.transitions = transition.map(el => new Transition(this, el));
+    
+
+    if (this.states && this.states.length){
+        const initial = this.initial || this.states[0].id;
+        this.push({id: initial});
 
 
-        if (this.states && this.states.length) {
-            const initial = this.initial || this.states[0].id;
-            this.push({id: initial});
+    }
+
+    }
+    }
+    export class BotEditPhotoState extends State {
+        get storage() { return this.parent.storage };
+        set storage(value) { this.parent.storage = value };
+        get xmppStore() { return this.parent.xmppStore };
+        set xmppStore(value) { this.parent.xmppStore = value };
+        get friendStore() { return this.parent.friendStore };
+        set friendStore(value) { this.parent.friendStore = value };
+        get profileStore() { return this.parent.profileStore };
+        set profileStore(value) { this.parent.profileStore = value };
+        get messageStore() { return this.parent.messageStore };
+        set messageStore(value) { this.parent.messageStore = value };
+        get searchStore() { return this.parent.searchStore };
+        set searchStore(value) { this.parent.searchStore = value };
+        get eventStore() { return this.parent.eventStore };
+        set eventStore(value) { this.parent.eventStore = value };
+        get model() { return this.parent.model };
+        set model(value) { this.parent.model = value };
+        get location() { return this.parent.location };
+        set location(value) { this.parent.location = value };
+        get pushStore() { return this.parent.pushStore };
+        set pushStore(value) { this.parent.pushStore = value };
+        get botStore() { return this.parent.botStore };
+        set botStore(value) { this.parent.botStore = value };
+
+    constructor(_, parent, sm){
+    super({ id: "BotEditPhoto"}, parent, sm);
+        const storage = this.storage;
+        const xmppStore = this.xmppStore;
+        const friendStore = this.friendStore;
+        const profileStore = this.profileStore;
+        const messageStore = this.messageStore;
+        const searchStore = this.searchStore;
+        const eventStore = this.eventStore;
+        const model = this.model;
+        const location = this.location;
+        const pushStore = this.pushStore;
+        const botStore = this.botStore;
+
+    let states = [];
+    let transition = [];
+
+    this.states = states;
+    this.transitions = transition.map(el => new Transition(this, el));
+    
+
+    if (this.states && this.states.length){
+        const initial = this.initial || this.states[0].id;
+        this.push({id: initial});
 
 
-        }
+    }
 
     }
     }
@@ -4505,17 +4503,13 @@ export class BotSubscriberListState extends State {
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.initial = 'Disconnected';
-        this.onentry = _event => {
-            this.sm.on({
-                $line: '263',
+    this.initial = 'Disconnected'; this.onentry = _event => {this.sm.on({$line: '266',
 $column: '29',
 $type: 'on',
 event: 'disconnected',
 content: () => {return xmppStore.disconnected}, 
 })
-            ;this.sm.on({
-                $line: '264',
+; this.sm.on({$line: '267',
 $column: '26',
 $type: 'on',
 event: 'connected',
@@ -4655,10 +4649,7 @@ content: () => {return xmppStore.connected},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.$type = 'parallel';
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '271',
+    this.$type = 'parallel'; this.onentry = _event => { this.sm.script({$line: '274',
 $column: '13',
 $type: 'script',
 content: () => {return this.model.load(_event.data)}, 
@@ -4720,18 +4711,12 @@ content: () => {return this.model.load(_event.data)},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '276',
+    this.onentry = _event => { this.sm.script({$line: '279',
 $column: '14',
 $type: 'script',
 content: () => {return friendStore.start()}, 
 })
-            ;
-        };
-        this.onexit = _event => {
-            this.sm.script({
-                $line: '279',
+; }; this.onexit = _event => { this.sm.script({$line: '282',
 $column: '14',
 $type: 'script',
 content: () => {return friendStore.finish()}, 
@@ -4790,18 +4775,12 @@ content: () => {return friendStore.finish()},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '289',
+    this.onentry = _event => { this.sm.script({$line: '292',
 $column: '14',
 $type: 'script',
 content: () => {return messageStore.start()}, 
 })
-            ;
-        };
-        this.onexit = _event => {
-            this.sm.script({
-                $line: '292',
+; }; this.onexit = _event => { this.sm.script({$line: '295',
 $column: '14',
 $type: 'script',
 content: () => {return messageStore.finish()}, 
@@ -4860,18 +4839,12 @@ content: () => {return messageStore.finish()},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '297',
+    this.onentry = _event => { this.sm.script({$line: '300',
 $column: '14',
 $type: 'script',
 content: () => {return location.start()}, 
 })
-            ;
-        };
-        this.onexit = _event => {
-            this.sm.script({
-                $line: '300',
+; }; this.onexit = _event => { this.sm.script({$line: '303',
 $column: '14',
 $type: 'script',
 content: () => {return location.finish()}, 
@@ -4930,18 +4903,12 @@ content: () => {return location.finish()},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '305',
+    this.onentry = _event => { this.sm.script({$line: '308',
 $column: '14',
 $type: 'script',
 content: () => {return eventStore.start()}, 
 })
-            ;
-        };
-        this.onexit = _event => {
-            this.sm.script({
-                $line: '308',
+; }; this.onexit = _event => { this.sm.script({$line: '311',
 $column: '14',
 $type: 'script',
 content: () => {return eventStore.finish()}, 
@@ -5054,9 +5021,7 @@ content: () => {return eventStore.finish()},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '315',
+    this.onentry = _event => { this.sm.script({$line: '318',
 $column: '14',
 $type: 'script',
 content: () => {return pushStore.start()}, 
@@ -5115,18 +5080,12 @@ content: () => {return pushStore.start()},
 
     this.states = states;
     this.transitions = transition.map(el => new Transition(this, el));
-        this.onentry = _event => {
-            this.sm.script({
-                $line: '320',
+    this.onentry = _event => { this.sm.script({$line: '323',
 $column: '14',
 $type: 'script',
 content: () => {return botStore.start()}, 
 })
-            ;
-        };
-        this.onexit = _event => {
-            this.sm.script({
-                $line: '323',
+; }; this.onexit = _event => { this.sm.script({$line: '326',
 $column: '14',
 $type: 'script',
 content: () => {return botStore.finish()}, 
@@ -5201,10 +5160,8 @@ export class Statem extends StateMachine {
     get botShareSelectFriends(): BotShareSelectFriendsState {return this.getState("BotShareSelectFriends")};
     get botInfo(): BotInfoState {return this.getState("BotInfo")};
     get botMap(): BotMapState {return this.getState("BotMap")};
-
-    get botSubscriberList(): BotSubscriberListState {
-        return this.getState("BotSubscriberList")
-    };
+    get botSubscriberList(): BotSubscriberListState {return this.getState("BotSubscriberList")};
+    get botEditPhoto(): BotEditPhotoState {return this.getState("BotEditPhoto")};
     get connection(): ConnectionState {return this.getState("Connection")};
     get disconnected(): DisconnectedState {return this.getState("Disconnected")};
     get connected(): ConnectedState {return this.getState("Connected")};