+++ src/components/BotPhotoList.js
@@ -147,19 +147,12 @@ BottomBar.defaultProps = {
 @autobind
 @observer
 export default class BotPhotoList extends React.Component {
-    constructor(props) {
-        super(props);
-        if (!bot.bot) {
-            bot.createImage();
-            bot.bot.addImage('tros:d6976ac8-5a3a-11e6-8008-0e2ac49618c7@staging.dev.tinyrobot.com/file/36d93122-a1ae-11e6-b428-0e600a8611a9', '1');
-            bot.bot.addImage('tros:d6976ac8-5a3a-11e6-8008-0e2ac49618c7@staging.dev.tinyrobot.com/file/36d93122-a1ae-11e6-b428-0e600a8611a9', '2');
-        }
-    }
 
     render() {
         if (!bot.bot) {
             return <Screen/>
         }
+        console.log("PHOTOS:", JSON.stringify(bot.bot.images));
         return <Screen>
             <Swiper style={styles.wrapper} height={height}
                     renderPagination={renderPagination}